Finish Lab1

+++ src/simpledb/HeapPage.java
@@ -79,7 +79,7 @@ public class HeapPage implements Page {
         @return the number of tuples on this page
     */
     private int getNumTuples() {
-        return (int)(Math.floor(BufferPool.getPageSize() * 8 / (td.getSize() * 8 + 1)));
+        return (int)(Math.floor(((double)BufferPool.getPageSize() * 8 / (td.getSize() * 8 + 1))));
     }
 
     /**
@@ -87,7 +87,7 @@ public class HeapPage implements Page {
      * @return the number of bytes in the header of a page in a HeapFile with each tuple occupying tupleSize bytes
      */
     private int getHeaderSize() {
-        return (int)(Math.ceil(getNumTuples() / 8));
+        return (int)(Math.ceil(((double)getNumTuples() / 8)));
     }
     
     /** Return a view of this page before it was modified
